Replaced more functions with nginx alternative. Separated out hardcoded strings from rest of find_request_length function.

+++ presentation_http_request.h
@@ -39,7 +39,7 @@ void presentation_http_request_close_connection(ngx_connection_t *c);
 
 /* Helper function to call the recv function with error checking */
 size_t recv_wrapper(ngx_connection_t *c, presentation_request_t *request, ngx_event_t *rev);
-int find_request_length(presentation_request_t *request);
+ssize_t find_request_length(presentation_request_t *request);
 
 
 #endif
